+++ src/game/objects/Player.java
@@ -231,7 +231,7 @@ public class Player extends Creature {
      * @param mapID
      * @param cellID 
      */
-    public void teleport(int mapID, int cellID){
+    public void teleport(short mapID, short cellID){
         if(GameMap.isValidDest(mapID, cellID)){
             MapEvents.onRemoveMap(session);
             MapEvents.onArrivedOnMap(session, mapID, cellID);
